requirements/base.txt
# Core Django
Django>=4.2,<5

# DRF
djangorestframework>=3.15

# JWT
djangorestframework-simplejwt>=5.3

# CORS
django-cors-headers>=4.3

# Env
python-dotenv>=1.0
django-environ>=0.12

# PostgreSQL driver
psycopg2-binary>=2.9

# Servidor WSGI (pode estar no production também)
gunicorn>=21.0

# Static files (para produção)
whitenoise>=6.5
